pyproject.toml
[project]
name = "algieba-llm-action"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
